+++ src/Client/ClientSetup.h
@@ -46,7 +46,7 @@ public:
         Window* speedhack = new Speedhack();
         speedhack->name = "Speedhack";
         speedhack->id = "speedhack-window";
-        speedhack->windowPos = ImVec2(50 + (50 + (Client::instance->tileSize.x)) * 2, 50);
+        speedhack->windowPos = ccp(50 + (50 + (Client::instance->tileSize.x)) * 2, 50);
 
         speedhack->modules.push_back(new SpeedhackTop());
         speedhack->modules.push_back(new SpeedhackEnabled());
@@ -64,7 +64,7 @@ public:
         Window* level = new Level();
         level->name = "Level";
         level->id = "level-window";
-        level->windowPos = ImVec2(50 + (50 + (Client::instance->tileSize.x)) * 1, 50);
+        level->windowPos = ccp(50 + (50 + (Client::instance->tileSize.x)) * 1, 50);
 
         level->modules.push_back(new Module("Noclip", "noclip", "Prevents the player from dying"));
         level->modules.push_back(new Module("Instant Complete", "instant", "Instantly completes a level.\nMay not work because of the <cg>Geometry Dash</c> anti-cheat.\nUsing this cheat on a <co>rated level</c> <cr>WILL GET YOU LEADERBOARD BANNED</c>", false, "<cr>Warning!</c>\nUsing this cheat on a <co>rated level</c> <cr>WILL GET YOU LEADERBOARD BANNED</c>", true));
@@ -135,7 +135,7 @@ public:
         Window* bypass = new Window();
         bypass->name = "Bypass";
         bypass->id = "bypass-window";
-        bypass->windowPos = ImVec2(50 + (50 + (Client::instance->tileSize.x)) * 0, 50);
+        bypass->windowPos = ccp(50 + (50 + (Client::instance->tileSize.x)) * 0, 50);
 
         //bypass->modules.push_back(new Module("Slider Limit Bypass", "slider-limit", "Allows sliders to go beyond the limit of the slider. <cr>Doesn't work for scaling in the editor currently</c>"));
         
@@ -147,7 +147,7 @@ public:
         Window* bypass = new Universal();
         bypass->name = "Universal";
         bypass->id = "universal-window";
-        bypass->windowPos = ImVec2(50 + (50 + (Client::instance->tileSize.x)) * 0, 50);
+        bypass->windowPos = ccp(50 + (50 + (Client::instance->tileSize.x)) * 0, 50);
 
 
         bypass->modules.push_back(new Module("Practice Music Sync", "practice-bypass", "Unlocks the Practice Music Sync button in the pause menu options"));
@@ -194,7 +194,7 @@ public:
         Window* creator = new Window();
         creator->name = "Creator";
         creator->id = "creator-window";
-        creator->windowPos = ImVec2(50 + (50 + (Client::instance->tileSize.x)) * 0, 50);
+        creator->windowPos = ccp(50 + (50 + (Client::instance->tileSize.x)) * 0, 50);
 
         creator->modules.push_back(new SetValueModule("Set Scale:", "set-scale"));
         creator->modules.push_back(new SetValueModule("Set Rotation:", "set-rot"));
@@ -232,7 +232,7 @@ public:
         Window* cosmetic = new Window();
         cosmetic->name = "Cosmetic";
         cosmetic->id = "cosmetic-window";
-        cosmetic->windowPos = ImVec2(50 + (50 + (Client::instance->tileSize.x)) * 3, 50);
+        cosmetic->windowPos = ccp(50 + (50 + (Client::instance->tileSize.x)) * 3, 50);
 
         //cosmetic->modules.push_back(new Module("Hide Endscreen BTN", "end-screen", "Adds an arrow to hide the end screen"));
         cosmetic->modules.push_back(new Module("No Transition", "no-trans", "Disables the fade scene transitions"));
@@ -293,7 +293,7 @@ public:
         Window* misc = new Window();
         misc->name = "Misc";
         misc->id = "misc-window";
-        misc->windowPos = ImVec2(50 + (50 + (Client::instance->tileSize.x)) * 4, 50);
+        misc->windowPos = ccp(50 + (50 + (Client::instance->tileSize.x)) * 4, 50);
 
         //misc->modules.push_back(new Module("Use Full Options", "full-options", "Opens the full options menu instead of the mini options in the pause menu.\nAlso adds a button to the normal pause menu to open the mini options."));
 
@@ -374,7 +374,7 @@ public:
         Window* replay = new Window();
         replay->name = "Config";
         replay->id = "options-window";
-        replay->windowPos = ImVec2(50 + (50 + (Client::instance->tileSize.x)) * 4, 50);
+        replay->windowPos = ccp(50 + (50 + (Client::instance->tileSize.x)) * 4, 50);
 
         if (android)
         {
@@ -394,7 +394,7 @@ public:
         Window* creator = new Window();
         creator->name = "Creator";
         creator->id = "creator-window";
-        creator->windowPos = ImVec2(50 + (50 + (Client::instance->tileSize.x)) * 4, 50);
+        creator->windowPos = ccp(50 + (50 + (Client::instance->tileSize.x)) * 4, 50);
 
         creator->modules.push_back(new Module("Copy Hack", "copy-hack"));
 
@@ -414,7 +414,7 @@ public:
         Window* theming = new Window();
         theming->name = "Theming";
         theming->id = "themes-window";
-        theming->windowPos = ImVec2(50 + (50 + (Client::instance->tileSize.x)) * 5, 50);
+        theming->windowPos = ccp(50 + (50 + (Client::instance->tileSize.x)) * 5, 50);
 
         theming->modules.push_back(new ColourModule("Accent", "accent-colour", ccc3(10, 0, 0)));
 
@@ -428,7 +428,7 @@ public:
         Window* devtools = new Window();
         devtools->name = "Developer";
         devtools->id = "devtools-window";
-        devtools->windowPos = ImVec2(50 + (50 + (Client::instance->tileSize.x)) * 5, 50);
+        devtools->windowPos = ccp(50 + (50 + (Client::instance->tileSize.x)) * 5, 50);
 
         devtools->modules.push_back(new Module("Recompile Shader", "recomp-shaders", "Recompiles blur shader"));