+++ src/Client/Module.cpp
@@ -173,8 +173,6 @@ void Module::enablePatches()
 
 void ColourDelegate::updateColor(cocos2d::ccColor4B const& color)
 {
-    log::info("colour: {}", color);
-
     auto v = as<ColourModule*>(mod);
 
     v->colour = ccc3(color.r, color.g, color.b);