+++ src/UI/PCDrawUtils.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <Geode/Geode.hpp>
+
+using namespace geode::prelude;
+
+namespace PCDrawUtils
+{
+    void drawRect(CCPoint pos, CCPoint size, ccColor4B colour);
+
+    CCPoint getMousePosition();
+};