+++ src/UI/PCGuiNode.hpp
@@ -1,9 +1,12 @@
-/*#pragma once
+#pragma once
 
 #include <Geode/Geode.hpp>
 #include <Geode/modify/CCEGLView.hpp>
 #include <Geode/modify/CCDirector.hpp>
+#include <Geode/modify/CCScheduler.hpp>
+#include <Geode/modify/CCTouchDispatcher.hpp>
 #include "../Client/Client.h"
+#include "../Utils/CCBlurLayer.hpp"
 
 using namespace geode::prelude;
 
@@ -15,12 +18,15 @@ class PCGuiNode : public CCNode
     public:
         int v = 0;
         bool drawScene = false;
+        CCBlurLayer* blur;
+        CCClippingNode* clip;
 
         static PCGuiNode* create();
         static PCGuiNode* get();
 
         virtual void visit();
         virtual void draw();
+        virtual void update(float delta);
 
         ~PCGuiNode();
-};*/
+};